On 64 bit use big-node-linkage for all subtree linkages.
Previously there was a test if (distance > 2^32) which was always true
due to confusing precedence rules.  Correcting this let to having a mixture of
normal linkage and big-node-linkage, which did not work correctly and was very
confusing.

This test and the 64 bit normal node linkage for subtrees has been removed.