optimized SimpleWaveData on MdsDataProvider
Stressed out the method of tdi-variable-caching. This way the the access
to the node is reduced to one time only compared to four if only Y-data
is provided (LocalDataProvider). This results in a major speed up if the
signal from the node is generated by a time-consuming method e.g. a
python web interface.
Some code clean up.
Could be intensified as so far I did not cross a case where v_idx was
not 0 consider replacing "_jscope_"+(var_idx  ) and
"_jscope_"+(var_idx+1) with _jscope_y and _jscope_x.