replace eval with *argin
I am not 100% this is there since the beginning but *arg is a much
better and saver way to parse inputs dynamically. The argument parsing
is a helper-function to be used by TdiCompile and TdiExecute. it turns
the tuple of arguments into a list of pointers. The *argin expands the
list into individual input arguments