add autostore method for use by boards connecting back and asking for store to be done
When checking for a trigger if the store method is being done by the user:
i.e. do /meth store
the code must wait for the board to complete its post processing step

When the board is doing a store, this is part of the post processing step, but performed
after the board has done its other 'system' post processing.  The store method used in this
case needs to count a 'I'm post processing' response from the board as OK