acq devices (and python devices in general) need to raise errors that actmon can display
When actions are dispatched if the error does not have a 'status' (ours do, but in general Execptions do not)
PyDoMethod.py sets the status to zero.

when actmon notices that a dispatched action has failed it trys to make a string from the action's status,
which is an integer filled in with the returned status from TreeDoMethod.

So acq.py was raising errors from Non MDSplus things, which were being displayed as 0 - NO MESSAGE For code 0
Now acq.py raises errors which have MDSplus status and strings so ACTMON should be more informative.