configure.ac: Remove bad characters from comments
The comments for the --enable-debug option were filled with unicode that git doesn't support.  Replace with standard characters for quotes.  In the case of the double dash being replaced by an em dash, in this case the double dash was correct, as it is exactly what the user will type at the command line.