javascope AboutWindow: Remove try/catch
The try/catch block in AboutWindow's ctor was swallowing any null
pointer exception that might have occurred.  Remove this, since there is
a debug print statement anyway, and if a null pointer exception occurs, we
want to notice and fix the bug.  Reindent the surrounding lines to
accomodate the change in scope level.