Kge javamds cleanup (#585)
* javamds/mdsobjects.c: Fix length type

The length argument should be a pointer to unsigned short, not signed short.  Adjust accordingly.

* javamds/javamds.c: Remove unused function

* javamds/JavaTrav.c: Fix invalid ptr use

If num_nids is zero, then nids is incorrectly dereferenced.  Instead, return early with NULL in this case to avoid the problem.  With the simpler logic, remove the extra if's and correct indentation due to the nesting change.

* javamds/javamds.c: Fix possible free of NULL pointer

expanded_in was being conditionally malloc'd and unconditionally free'd.  Change to be both unconditionally malloc'd and free'd.  Remove the now-redundant free's, and switch to calloc to avoid any potential for string side effects with the new scoping.