javascope/MdsConnection: Fix error string concatenation
In the Disconnect and Quit functions, the String::concat function return
value was not properly being used.  Strings are immutable, so the call
to str.concat("xxx") will not modify the str variable.  It will instead
return a new String whose contents are the concatenation of str and
"xxx".  So in these cases, the error string was never actually being
modified with the intended error text.  Fix by replacing all of this
with simple string concatenation operators (str + "xxx").