config: Only require specific java versions for mdsobjects
We only need to restrict Java environments for the mdsobjects interface,
and only because of older (<=2014) versions of MatLab having a built in
JRE6.  So, to allow jScope and jTraverser to build with (and use
features of) newer versions, only apply the -source, etc. options to the
mdsobjects/java/Makefile.am build.