replace PY_DEV_NOTFOUND2 with PY_INTERFACE_LIBRARY_NOT_FOUND
The original PY_DEV_NOTFOUND is used to signal that support for a given
MDSplus device type is not available in python.  This error code was
reused by the picam (princeton instruments camera) module, when the
PiCam python interface to the princeton instruments library was unavailable.

At some point the it was rename to have a 2 at the end.

This change reuses the same error value and gives it a new more meaningful name.