mdsobjects.h: Fix pointer casting
In the Compound ctor, the ptr argument was a char *, and used with C-style casts to convert to different types.  This is not correct for C++.  Change the argument to void * and replace with static_cast, which is well defined for the language.  This does not guarantee that the user hasn't made an error, but it at least guarantees that the conversions are correct.