mdsobjects.h: Only set clazz once
Compound has multiple constructors that all set clazz.  In C++11, we could handle this more gracefully.  But, to support C++03, at least we can factor this out to a setClass member function and call that from each constructor in the same place.  To that end, it makes sense then to gather all of the base class initializations in the same place.