Use symbolic links for mdsip plugins which use tunnel protocol
The make of mdstcpip created symlinks for plugins like libMdsIpSSH.so
which point to libMdsIpTunnel.so. This feature tunnels mdsip messaging
through shell scripts which can be used for setting up communications
between the client and server. These symlinks were replaced by copies
of the shared library when performing make install because the
/usr/bin/install application was used to install these libraries in
the installation root. The rules for installing these tunneling libraries
now create symlinks instead of using /usr/bin/install.