Fix buffer overrun in tdi trace messages with compiling builtins
When tdishr attempts to compile builtins using the builtin names
such as add(), subtract(), etc... it miscalculates in the yacc
code the exact location of the error if these are called with
zero arguments and then writes past the end of allocated memory.
This fix should prevent this type of memory overrun.