New feature:: add MDSIP_CONNECT_TIMEOUT environment variable
The mdstcl dispatch commands potentially hang when trying to connect
to servers that are on the network but not responsive.

MdsIpShr has a connect_timeout parameter
set with:
  SetMdsConnectTimeout(int sec)
and retrieved with:
  GetMdsConnectTimeout()

This change uses the environment variable MDSIP_CONNECT_TIMEOUT to
set the default value for this parameter.  If it is undefined then
the original default of 0 is used.

Set this in envsyms with:
  MDSIP_CONNECT_TIMEOUT 1

or from the shell
  export MDSIP_CONNECT_TIMEOUT=1