Build: Update configure to define new MDSPLUS_SEARCH_LIBS function
The first attempt to ensure the TARGET_ARCH compiler options are
applied when checking for libraries was implemented by redefining
AC_CHECK_LIB and AC_SEARCH_LIBS. This change performs the same action
but does so without modifying the existing functions. It also replaces
the use of AC_CHECK_LIB with AC_SEARCH_LIBS which is the recommended
approach. The overall effect is to ensure that the checks used to
check for the existance of libraries do so using the architecture (32-bit
or 64-bit) the build is being performed for.