Build: Better sanitize and valgrind build option handling
The sanitize and valgrind tests supported by the operating system
and the docker used to build for that operating system are specified
in the os/*.opts files. Before this change, these os opts were overriding
the test selections so there was no way to skip these tests. This change
requires the --sanitize (with no test list) or --valgrind options to be
set or none of the tests will be executed. If sanitize or valgrind tests
are desired the trigger.sh and/or build.sh scripts should contain these
options without test lists to enable the tests.