Build: Add distclean target to top Makefile
This commit adds the distclean target to the top directory Makefile.
The command "make distclean" should clean the directory tree of any
files or directory not in the original source tree. If the sources
where obtained via git it will use the "git clean" command to remove
the files otherwise it will use "make fullclean" if the "make distclean"
command was issued from the top source directory. It is now recommended
that configure and make be performed from a build directory outside
of the source tree which should leave the source directory tree in
a clean state.

This commit is to address git issue:

https://github.com/MDSplus/mdsplus/issues/680