Build: Remove --key-id option on rpmsign
When adding the new RSA signing keys for debian packages the --key-id option was
added to the redhat rpmsign command to ensure the original key was used to sign
the rpms. It turns out that the --key-id option is not supported on redhat distributions
earlier than rhel7 and the switch to use different directories to separate the signing
keys removed the need to use the --key-id option.