Build: Fix tests when building alpine distributions
A few minor changes were required to get the standard tests to work on
the alpine linux system. There is a problem with the mdsip udt plugin when
running on the armhf architecture so the udt test is now disabled. It is very unlikely
that udt would ever be used on this architecture so disabling the test should be ok.
Some of the existing tests rely on the type char to be a signed number and it appears that on the armhf architecture it is defaulting to unsigned so an -fsigned-char option is necessary.