Build: Fix cleanup of mdsplus python modules during upgrades
The postinstall script used to remove the older mdsplus python modules
when updating the python rpm was using an mdsplus* wildcard in the shell
command which works fine unless there is a file/directory in the / directory
where the script is executed. The mdsplus* string has to be quoted otherwise
the shell does the substitution too early instead of applying the wildcard in
the find command. Subtle!