FIX: TreeWriteTree was not returning an error when output disk full
The error_exit code in TreeWriteTree  did not set status before returning.
When the code detected an error writing to the tree file, it returned, but
status was 1.  So the calling application thought the write operation was
successfull

Added new tree error code TreeWRITETREEERR  to be returned under these
conditions.