Build: Split alpine builds into 3 separate os so they can be built in parallel
Unlike the debian repositories the alpine linux repository has a separate
repository contents file for each of the architectures so the installers and
repository can be built independently of each other. The armhf build can now
be built on a rasperyPi configued as a jenkis slave separately from the x86
and x86_64 architecture builds. To avoid conflicts in the files used to tag
when a release has been built the build.sh script was changed to include the
os name in the tagfile name (i.e. ${PUBLISH}/${DISNAME}/${BRANCH}_${VERSION}_${OS}.