Fix: Use prefix on all lex processing to prevent conflicts with global symbols
When using lex/flex in constructing parses it will generate global symbols such as yytext.
Since a couple of the MDSplus libraries use lex this seemed to cause a problem on platforms
such as alpine. Adding a -P <prefix> to the lex command will generate code wich unique
prefixes to prevent conflicts.