FIX: - show server command after a stop server command crashes
doing double free

After sending a stop server command, the 'read answer info' call returns an
error, which is fine, since the server does not answer, however the error
case was only doing perror, and then falling through to the free of any
allocated memory.  Since the 'job' never completes, the job is still in the
que.  When the next 'show server' command comes along free(malloc) chokes
because it is trying to free memory that is not allocated.