Feature: Add command line recall to tditest
This change adds command line recall to the tditest utility.
If tditest is executed without any arguments it will save expressions
it receives and successfully evaluates in a history file. The command
line recall supports readline keystroke macros for doing things like
searching for a command in the history etc. Also  the echo of the comm
ands entered from stdin are no longer echoed.