Fix: Prevent python race condition deadlock in tree ctx handling
Previous __del__() method for tree context acquired a lock before
modifying tree ctx tables. Unfortunately the python garbage collector
can execute at any time and if the lock happened to be already locked
when garbage collection begins this causes a deadlock which will cause
python to hang until the process is killed. This fix performs the operations
that require locking in a separate thread preventing this deadlock
condition.