Fix: invalid read in mdsdcl on env name
Before a memory corruption occured as value would point beyond the allocated memory of teh input string.
Now, if '=' is not in the P1 parameter, i.e. value[0]=='\0' after the for-loop, the current value of the env <name> is returned instead. This feature does not break any former use of mdsdcl_env.