Build: Use shell environment variable to specify OpenMotif dir on MacOSX
Older versions of MacOSX (i.e. Mountain Lion) installed the OpenMotif
development package in the /usr directory and this was hardcoded in
the configure script. Newer versions install this
in /usr/local and prevent the use of symlinks in /usr as a workaround.
This change enables the use of an OPENMOTIF environment variable to
specify the location of the OpenMotif headers and libraries.

Also moved -Wno-clobbered option to os specific settings of CFLAGS
since the clang compiler on MacOSX does not support that option.