Build: Use tmpdir for windows testing winebottle
When tests are performed on the Windows platform a winebottle
was being created in the current directory for tests to be performed
using wine (windows emulation on linux). This winebottle has symlinks
to the root filesystem and this causes problems when doing builds
using jenkins. The jenkins build searches for various types of files
in the workspace directory tree and if the winebottle is in the
workspace directory these symlinks cause an infinite recursion. This
PR will cause docker builds for windows to create the winebottle in
a subdirectory of /tmp to avoid this recusion problem.