Feature: Add username, date, and connection info to mdsip-ssh.log
When a user connects to a server using the ssh protocol, the server
runs mdsip-ssh-server and directs stderr to ~/mdsip-ssh.log.

This change adds information about the connection to the file. There
does not appear to be a sample server script for windows, only a client
script, so this only applies to linux like platforms.

Incomming ssh connections to windows servers have not been tested or
implemented.