Build: Fix email address used to send bad commit message title
The commit message title (first line) is used to determine the type of version
increment to use. This title syntax is checked and an email is sent to the author of the commit if the commit message does not follow our standard format. This fix is to use the email of the author of the commit not the account running the build job.