FIX: Connection timeout
A timeout argument has recently been added to mdstcpip/GetAnswerInfoTS(). However this routine was also used in
Connection C++ and java objects. Since there is no header in mdstcpip, the routine signature was manually reported in those libraries, and therefore the timeout was not reported and a random value was passed as timeout. This is likely also the cause for the random failures in Connection tests during build.