Fix: Add timeout argument to calls to GetAnswerInfoTS
A timeout argument was added to the GetAnswerInfoTS routine but most of the calls to
that routine were modified to include this argument. This provides a timeout of 0
for those calls which disables timeouts on the calls (the original behavior). We may
want to consider enabling timeouts in certain instance where appropriate later.