Gm fix connection timeout (#1047)
* FIX: Connection timeout

A timeout argument has recently been added to mdstcpip/GetAnswerInfoTS(). However this routine was also used in
Connection C++ and java objects. Since there is no header in mdstcpip, the routine signature was manually reported in those libraries, and therefore the timeout was not reported and a random value was passed as timeout. This is likely also the cause for the random failures in Connection tests during build.

* FIX: removed debug message

removed a forgotten debug message

* FIX: streaming timeout

Added a timeout when waiting new data for server side stream management in order to give a chance to xinetd mdsip processes to exit when the connection is no more active.

* Fix: Add timeout argument to calls to GetAnswerInfoTS

A timeout argument was added to the GetAnswerInfoTS routine but most of the calls to
that routine were modified to include this argument. This provides a timeout of 0
for those calls which disables timeouts on the calls (the original behavior). We may
want to consider enabling timeouts in certain instance where appropriate later.