Fix: Repair a bug in the mdsip timeout handling
The existing code called setsockopt with a 0 for the option
value when attempting to clear the socket timeout but that does
not work and setsockopt returns -1 and EINVAL for error so the
timeout does not get cleared. To clear the timeout you have to
send a timeout value (sec=0,usec=0) for the option value instead.