Build: Allow gdb use inside docker builds (#1061)
When a build hangs it may be useful to run gdb on the process inside
the docker container to determine the cause of the hang. The SYS_PTRACE
capability must be enabled before you can use gdb inside the container.
To gdb the process you need to use "docker exec" to run gdb in the container.