Feature: Thin Client TreeNode object
A new class TreeNodeThinClient has ben added. It inherits from TreeNode and is instantiated bu Connection Method getNode("Path"), exporting TreeNode functionality over thin client mdsip.
There are several limitations in respect of the original TreeNode:

1) No Tree context is defined, Therefore a TreeNodeThinClient instance must be created after calling Connection.openTree() and is valid until the tree is closed or a new one is open  (remotely);

2)A subset of methods is implemented. The others generate an exception. Methods not supported include tree editing and derivation of relative nodes. Despite this, in practice all the normal use cases are covered.

3)Since thin connection is used, there may be some limitation in expression definition. For example, the dimension passed to method makeSegment() may be a Range type, but it is converted to an array before passing it to the mdsip server. Users must be aware of this.

Since ThreeNodeThinClient is a subclass of TreeNode, users are not aware of it. Therefore Connection.getNode() is declared to return a TreeNode instance.