Feature: Provide WRITE_ONCE enforcement with segmented signals (#1068)
* Feature: Provide WRITE_ONCE enforcement with segmented signals

If a node is set to be WRITE_ONCE (ie. TCL> SET NODE sig/WRITE_ONCE)
you will not be permitted to add or update any segments unless the
node is empty at which point only one segment can be written before
writes are blocked.

This also fixes the length characteristic of segmented nodes. The
length characteristic is a 32-bit value so the length value will
peak around 4 billion even if the total length of all the segments exceeds
this amount.

* Fix: nci needs to be unlocked

* Fix: Change max displayed length to max value for uint32_t

This changes the maximum length attribute of a segmented node to
4294967295, the maximum value of a uint32_t. It also changed the
data types returned for length and rlength by getnci to an
unsigned long instead of a long to prevent a negative length value.