Fix: Clear write once attribute when writing segments into an empty node
The write_once attribute when used with segments is only meaningful when
set after the completion of intended storage of all segments. If a node
has the write_once attribute set before storage begins it will only permit
one segment write to the node which will break many existing devices.

This fix will clear the write_once attribute on a node if the node contains
no data when the first segment is written.