Fix: Remove deprecated key handling mechanisms
Java 1.4 added new ways to access key presses in an event handler.  Java
9 deprecated the old way.  This basically involves just changing X_MASK
to X_DOWN_MASK, and any use of Event.X_MASK to the actual type of event,
such as MouseEvent.  In the future, my guess is that this repetitive
code should be factored out into a local private method that checks for
the keypress.