Build: Add script for notifying developers of issues via email
This commit adds a python script for use in a scheduled jenkins job
which will send email notification to the MDSplus developers of any
outstanding issues. The email will contain a list of the oustanding
issues that have been assigned to the user followed by a list of
outstanding issues that have yet to be assigned to anyone.

NOTE: If an issue is given a label "noremind" it will not be included
in these issue notification emails.

This is to provide the enhancment described in
https://github.com/MDSplus/mdsplus/issues/1076