Feature: Add ability to "import xyzDevices"
The original implementation of MDSPLUS_DIR/pydevices prevented the
ability of importing the xyzDevices (.e. MitDevices) module assuming
the device support modules would only be referenced when adding or
operating an instance of a particular device. This change restores the
ability of doing: >>> import MitDevices or >>> from MitDevices import mydevice.