Build: Remove new compiler warnings introduced in GNUC 7
Version 7 of the GNU compilers now warn about case statements not
including a break and therefore falling through to the following
case statement. This is a legal and common practice but the new
warning is to avoid common bugs in software where the fallthrough
was unintended because the break statement was forgotten. The code
must now explicitly indicate that the fallthrough was intentional.

There is a caveat to this... compilers earlier than version 7 cannot
parse the attribute command used to ignore the fallthrough warning.
This has been implemented as a macro, MDSPLUS_ATTR_FALLTHROUGH,
in <mdsplus/mdsplus.h>. When using pre-version 7 compilers this is
defined as an empty macro. When using >= version 7 compilers this will
be defined as: __attribute__ ((fallthrough));

In a case statement that should intentionally fallthrough to the next
case you will need to do the following:

 #include <mdsplus/mdsplus.h>
    ...
    case xxx:
      do something with no break or return
      MDSPLUS_ATTR_FALLTHROUGH
    case yyy:
    ...