Build: User real source directory inside docker images
The original docker build scripts used /source for the source volume.
Various development tools can be used better if the docker builds use
the same source directory spec as that outside of docker. For example
if the source lives in /home/twf/mdsplus-alpha and that same directory
is used as the internal volume inside the docker container instead of
mapping it to /source, all debug information etc will contain the correct
source file location if used outside of docker.