Build: add --docker-srcdir option to build and stop using /sources dir in build.sh
This change will add a new --docker-srcdir option for specifying the source
directory using in docker containers that are run by build.sh. It
also changes the default source directory used inside docker from /source
to the actually directory specification of the sources used on the host.

While making the necessary changes to the windows mdsplus.nsi file having
to deal with the removal of the hardcoded /source directory, the url specified
for downloading the sources in the windows installer was fixed to use github
to download a source zip file.