Fix: Fix _TreeRenameNode (#1119)
The _TreeRenameNode code erroneously made calls to TreeXXX routines instead
of the _TreeXXX version of the routines passing a dbid. This caused problems
in usage in MdsObjects because these calls attempted to perform the function
using the global tree ctx which is wrong. The problem was detected by valgrind
because the TreeGetNci call was actually failing so the variables it was supposed
to be initializing remained uninitialized. There were more instances of misuse
of the TreeXXX variant of the call which this fix also corrects.