Fix: convert argument to ctypes c_int32() constructor to int
When constructing ctype data instances such as c_int32 it is necessary
to use a python integer as the argument or it produces an error. Using
python bridges like that provided by matlab which defaults to creating
floating point values for integer constants things like shot numbers and
segment indexes must be converted to python integers to avoid errors
constructing the c_int32 instances used for calling the MDSplus libraries.
This change explicitly converts the arguments to integers before constructing
the c_int32 instances.