Fix: Restore MDSplus python exception superclasses (#1127)
* Fix: Restore MDSplus python exception superclasses

The MDSplus python exception superclasses were changed to begin
with an underscore to prevent them from being exported but it is
sometimes useful to catch a class of exceptions by using the superclass
name. Some existing applications were doing this and the change to
hide these superclasses broke these applications. This change will
restore their visibility making them useable in user applications.

* Fix: Repair duplicate MDSplusException definitions

Removing the underscore in the exception superclasses caused
a problem with the MDSplusException class being redefined. This
fixes that problem.