Fix: Repair installer scripts again
The problem with the current scripts is due to the unexpected order
of the execution of the scripts when installing, upgrading or removing
rpm packages. It turns out that the uninstall scripts are performed
after the install scripts. The uninstall scripts were attempting to
delete the pyc modules in the pydevice subdirectories by removing
the entire subdirectory. Since on redhat systems the uninstall scripts
run after the install scripts this was removing the python device directory
just installed. On redhat systems the scripts are called with an argument
indicating if the package is being removed or updated. This argument will
be used to avoid deleting the directory if this operation was an update.