Build: Turn on --enable-werror for release, optimized, build (#1154)
* Build: Turn on --enable-werror for release, optimized, build

This change will enable -Werror for all builds including release
builds with -O3 compiler optmization. Some platforms are generating
unused result warning when certain functions are used and the return
value is not used. At some point we should investigate these warnings
and properly check the return values and take the appropriate action
if the return is not what is expected. For now, this change also
adds a --disable-wreturns configure option and the builds use it so that
these warnings will ignored.

* Build: Use __attribute__((alias(name)) for Fortran entry points

Instead of defining separate fortran entry points this change will use
the __attribute__ feature of gcc to define the Fortran entry points as
aliases to the C routines. This simplifies mdslib.c and removes
compiler warnings.

* Build: mdslib aliases must be weak on OSX

* Build: Remove curl write callback compiler warning

* Build: Repare apple fix typo

* Build: Another attempt to stop warnings on macosx and raspberrypi

* Build: Fix false positive warning in javamds.c

On raspberrypi a warning occurs about the possibility of the variable
jxDouble being used before being initialized. I don't understand how
that might happen. To remove the warning I am initializing jxDouble
to NULL in the variable declaration.