Fix: Repair rpm repo public signing key
The signing key for MDSplus has been changed to use a more secure
key. This was fixed because of warnings occurring for debian repositories
but the public key installed for the rpm repository was not updated to
match the new signing key. This should cause a new public key to be
included in the repo rpm.