Fix: Prevent hangs during removal of python packages
The installer scripts were changed to use pip to remove old python
packages left from previous installations. Unfortunately pip prompts
for yes/no verification and this hung the removal when done using the
package install/remove scripts which are not done interactively. This
should fix this problem. Unfortunately to update any of the python
packages it may be necessary to manually edit the mdsplus-python.prerm
script found in /var/lib/dpkg/info on debian based linux systems
to change the line:

 while (pip uninstall -q -y MDSplus 2>/dev/null);do :;done

to be:

 while (/usr/bin/yes | pip uninstall -q MDSplus 2>/dev/null);do :;done

before an update or removal can be performed.