Feature:handle names longer than 12 chars
Node names longer that 12 chars are now splitted and a subtree created. For example, adding a node whose name 'THISISALONGNODENAME' will ceate a subtree whose root is names 'THISISALONGN' and a son names 'ODENAME'. This feature is intended for usage in ITER IMAS where a way for mapping long names into tree nodes is required.