Tests: Add tditest testing
This commit adds a set of tests driven by files containing
simply tdi expressions. The tdi expressions are processed by the
tditest program and the output (both stdout and stderr) are checked
to match an expected output. New tests can be added to existing
files of tdi expressions or new files can be added. If new test files
are added the tests.list file must be modified to include the new
tests. When adding new tests you need to invoke the do_tditest.sh
script passing update as an argument and it will regenerate the
test-name.ans files which then become the expected results.