Fix: Fix problem with adding extended attributes to empty nodes
The code for adding extended attributes mishandled nodes that
did not contain data. Instead of writing the attributes header
to the end of the datafile it instead wrote it to beginning
of the datafile. Subsquent adding of attributes to other nodes
without data also wrote it to the beginning of the datafile
overwriting any other settings found in that location. This fix
ensures that if there are no extended attributes defined for a node
the extended attributes will be written to the end of the datafile.

Tests for TCL> SET/SHOW ATTRIBUTE commands were also added to the
tditest/testing/tdi-tcl.tdi test.