Feature: Enable manipulation of node extended atts in python
This adds the capability to manipulate extended attributes of
nodes in MDSplus tree using the python MDSplus module.

A node attribute can be added or modified by using the
setExtendedAttribute(name,value) method of a TreeNode.
The value of an extended attribute can be fetched using
the getExtendedAttribute(name) method. A python dictionary
of all extended attributes of a node can be obtained using
the getExtendedAttributes() method. Some special node
property shortcuts can be used as well. A TreeNode
instances property whose name begins with two underscores
can be used to set or get an extended attribute.

The following demonstrates these new features:

>>> t=Tree('main',1)
>>> t.NUMERIC.setExtendedAttribute('att1',42)
>>> t.NUMERIC.setExtendedAttribute('att2',Range(1,100))
>>> t.NUMERIC.getExtendedAttribute('att1')
42
>>> t.NUMERIC.getExtendedAttribute('att2')
1 : 100 : *
>>> t.NUMERIC.getExtendedAttributes()
{'att2   ': 1 : 100 : *, 'att1   ': 42, 'special': 42}
>>> t.NUMERIC.__att3=100
>>> t.NUMERIC.__att2
1 : 100 : *
>>> t.NUMERIC.getExtendedAttributes()
{'special': 42, 'att3   ': 100, 'att1   ': 42, 'att2   ': 1 : 100 : *}
>>> t.NUMERIC.__special=None
>>> t.NUMERIC.getExtendedAttributes()
{'att3   ': 100, 'att1   ': 42, 'att2   ': 1 : 100 : *}