Fix: Sibling_of not handing members of usage structure correctly
sibling_of was assuming that nodes of usage structure must be
children. This is the case for nodes added with python, though using
other APIs it is evidently possible to have a member of usage structure.

This change makes sibling_of explictly check for member/child ness of
nodes  when assembing all the members and children of a node.