Fix: Handling of segments which have expression dim
The attempt to store segments which contain expressions for
the dimension of a segment uncovered problems in the handling
of segments. The existing code assumed that the dimensions of
segments were either ranges or arrays when combining multiple
segments into a single signal. This PR should fix some of the problems.
There is still further work needed to either support other such
features and/or restrict what can be stored in a segment to
prevent unexpected behaviors.

This should fix https://github.com/MDSplus/mdsplus/issues/1291