Fix:  wildcard searches for large trees very slow
The wildcard search operations
   ~~~
   :::
   ...
for very large trees was slow because the code needed to find the
end of the lists every time an element was added.

This PR keeps track of the end of NODELISTS being created so the lists
do not have to be traversed in order to append to them.